+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fixes Single-click into the window accidentally starting a selection.
 - Fixes font family change in live config reload.
 - Fixes config `initial_working_directory`'s' ~ (tilde) expansion failing to chdir, if followed by a `/`.
+- Adds scrollbar and configuration for scrollbar.
 
 #### prerelease 3